fix(comm): change timeout constants to use uint16_t type

+++ src/libcommon/comm.h
@@ -20,9 +20,9 @@
 #include <cstdint>
 #include <string>
 
-inline constexpr uint8_t COMM_SHORT_TIMEOUT = 1000;
-inline constexpr uint8_t COMM_AVERAGE_TIMEOUT = 10000;
-inline constexpr uint8_t COMM_LONG_TIMEOUT = 60000;
+inline constexpr uint16_t COMM_SHORT_TIMEOUT = 1000;
+inline constexpr uint16_t COMM_AVERAGE_TIMEOUT = 10000;
+inline constexpr uint16_t COMM_LONG_TIMEOUT = 60000;
 
 inline constexpr char MSG_TYPE[] = "type";
